rag_poc/requirements.txt
# Core web
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0

# RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain-text-splitters>=0.3.0

# HTTP client (for parallel calls)
httpx>=0.27.0

# LLM (reuse mvp deps - these are also needed for api)
openai>=1.0.0
